问题在备注中，加油吧骚年！~！~

+++ ZFB/Class/Main/Order/Model/SureOrderModel.h
@@ -13,7 +13,7 @@
 
 @property (nonatomic, strong) NSArray<Storedeliveryfeelist *> *storeDeliveryfeeList;
 
-@property (nonatomic, assign) NSInteger goodsCount;
+@property (nonatomic, assign) CGFloat goodsCount;
 
 @property (nonatomic, assign) CGFloat costNum;
